Gjoshi/rebase from master (#18)
Support for webhook event type and client refactoring

We need to check if the remote Bitbucket Server supports specific event
type before registering a webhook. The Factory client tests are
refactored so that multiple URL and their response can be mocked.

Also, Client factory should only be responsible for creating various client
and should delegate all HTTP execution-related logic to a different
module. Bitbucket client factory and all the other classes in client
package should not have a source code dependency on Jenkins related classes. 
Moving credentials to a package help in breaking this dependency
and moving all logic related to credential handling in one place. One of the
positive side effects of this refactoring is readable test classes.