Fix leaking file descriptor (#31)
File was opened but not closed. A new file was opened every time the trigger ran leading to a leak of file descriptors. Depending on host OS this could be a problem. Windows tests failed due to this (which prompted the finding). Switching to a try/with resource to automatically close the file when we're done with it.