Dk/add multibranch support (#105)
Add support for Multibranch projects. 

This adds support for multibranch pipelines. As such there is not a whole lot of new functionality, but a lot of code that is very similar to existing code. It adds a whole new set of files and UI code for BitbucketSCMSource. The SCM source is very similar to existing code and incorporates much of the same logic.